pyproject.toml
[project]
name = "xai-for-stress-and-mental-health-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdown>=5.2.1",
    "jupyterlab>=4.5.5",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]
